requirements.txt
awscrt==0.21.1
awsiot==0.1.3
AWSIoTPythonSDK==1.5.4
awsiotsdk==1.22.0
boto3==1.35.59
botocore==1.35.59
click==8.1.7
jmespath==1.0.1
ninja==1.11.1.1
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.3
six==1.16.0
tzdata==2024.2
urllib3==1.26.20
